feat:헤더 z-index 수정 ㄴsginup css디자인 수정 버그 수정
