+++ app/(main)/store/sign-up/components/GuildeText/GuideText.stories.tsx
