+++ app/(main)/store/sign-up/components/Input/Input.tsx
@@ -7,8 +7,8 @@ const Input = ({
 }) => {
   return (
     <label>
-      <div>{nameProp}</div>
-      <div className="h-12 rounded-l  border-white bg-white">
+      <div className="text-xs">{nameProp}</div>
+      <div className="h-12  rounded-l  border-white bg-white">
         <input
           {...props}
           className="h-12 w-full rounded-l border-white bg-white"
