+++ app/(main)/store/sign-up/components/NicknameInput/NicknameInput.stories.tsx
