+++ app/(main)/store/sign-up/components/NicknameInput/NicknameInput.tsx
@@ -14,12 +14,12 @@ const NicknameInput = ({
   };
   return (
     <label>
-      <div>{nameProp}</div>
+      <div className="text-xs">{nameProp}</div>
       <div className="flex flex-row ">
         <div className="h-12 w-9/12 rounded-l border-white bg-white">
           <input
             {...registerProp}
-            className="h-12 rounded-l border-white bg-white"
+            className="h-12 w-full rounded-l border-white bg-white"
           />
         </div>
         <div className=" ml-1 w-3/12">
