+++ package.json
@@ -68,7 +68,7 @@
   "lint-staged": {
     "*.{js,jsx,ts,tsx}": [
       "prettier --write",
-      "eslint --fix --max-warnings=0"
+      "eslint --fix --max-warnings=1"
     ]
   },
   "msw": {
